Standalone Stack Installers • Allow Custom Stacks
