+++ src/Contracts/StackInstaller.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Laravel\Breeze\Contracts;
+
+interface StackInstaller
+{
+    /**
+     * Install the Breeze stack.
+     *
+     * @return void
+     */
+    public function install(): void;
+}
